Extend parameterized file lists to Msc makefiles
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@1506 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/wc3270/gdi_print.c
@@ -572,7 +572,7 @@ gdi_init(const char *printer_name, const char **fail)
      */
     if (uparm.font_size) {
 	/* User-specified fixed font size. */
-	fheight = uparm.font_size * pstate.yptscale;
+	fheight = (int)(uparm.font_size * pstate.yptscale);
 	fwidth = 0;
     } else {
 	if (uparm.spp > 1) {
